+++ common/buildcraft/factory/TileQuarry.java
@@ -233,7 +233,7 @@ public class TileQuarry extends TileBuildCraft implements IMachine, IPowerRecept
 		if (!findTarget(true)) {
 			// I believe the issue is box going null becuase of bad chunkloader positioning
 			if (arm != null && box != null) {
-				setTarget(box.xMin + 1, yCoord + 2, box.zMin + 1);
+				setTarget((int) box.xMin + 1, yCoord + 2, (int) box.zMin + 1);
 			}
 
 			isDigging = false;
@@ -330,7 +330,7 @@ public class TileQuarry extends TileBuildCraft implements IMachine, IPowerRecept
 				for (int searchZ = startZ; searchZ != endZ; searchZ += incZ) {
 					if (!blockedColumns[searchX][searchZ]) {
 						Integer height = columnHeights[searchX][searchZ];
-						int bx = box.xMin + searchX + 1, by = searchY, bz = box.zMin + searchZ + 1;
+						int bx = (int) box.xMin + searchX + 1, by = searchY, bz = (int) box.zMin + searchZ + 1;
 
 						if (height == null) {
 							columnHeights[searchX][searchZ] = height = worldObj.getHeightValue(bx, bz);
@@ -571,8 +571,8 @@ public class TileQuarry extends TileBuildCraft implements IMachine, IPowerRecept
 			useDefault = true;
 		}
 
-		int xSize = a.xMax() - a.xMin() + 1;
-		int zSize = a.zMax() - a.zMin() + 1;
+		int xSize = (int) a.xMax() - (int) a.xMin() + 1;
+		int zSize = (int) a.zMax() - (int) a.zMin() + 1;
 
 		if (xSize < 3 || zSize < 3 || ((xSize * zSize) >> 8) >= chunkTicket.getMaxChunkListDepth()) {
 			if (placedBy != null) {
@@ -587,9 +587,9 @@ public class TileQuarry extends TileBuildCraft implements IMachine, IPowerRecept
 			useDefault = true;
 		}
 
-		xSize = a.xMax() - a.xMin() + 1;
-		int ySize = a.yMax() - a.yMin() + 1;
-		zSize = a.zMax() - a.zMin() + 1;
+		xSize =(int) a.xMax() - (int) a.xMin() + 1;
+		int ySize = (int) a.yMax() - (int) a.yMin() + 1;
+		zSize = (int) a.zMax() - (int) a.zMin() + 1;
 
 		box.initialize(a);
 
@@ -631,17 +631,17 @@ public class TileQuarry extends TileBuildCraft implements IMachine, IPowerRecept
 	}
 
 	private void initializeBlueprintBuilder() {
-		Blueprint blueprint = Blueprint.create(box.sizeX(), box.sizeY(), box.sizeZ());
+		Blueprint blueprint = Blueprint.create((int) box.sizeX(), (int) box.sizeY(), (int) box.sizeZ());
 
 		for (int it = 0; it < 2; it++) {
 			for (int i = 0; i < blueprint.sizeX; ++i) {
-				blueprint.setSchematic(worldObj, i, it * (box.sizeY() - 1), 0, BuildCraftFactory.frameBlock, 0);
-				blueprint.setSchematic(worldObj, i, it * (box.sizeY() - 1), blueprint.sizeZ - 1, BuildCraftFactory.frameBlock, 0);
+				blueprint.setSchematic(worldObj, i, it * ((int) box.sizeY() - 1), 0, BuildCraftFactory.frameBlock, 0);
+				blueprint.setSchematic(worldObj, i, it * ((int) box.sizeY() - 1), blueprint.sizeZ - 1, BuildCraftFactory.frameBlock, 0);
 			}
 
 			for (int k = 0; k < blueprint.sizeZ; ++k) {
-				blueprint.setSchematic(worldObj, 0, it * (box.sizeY() - 1), k, BuildCraftFactory.frameBlock, 0);
-				blueprint.setSchematic(worldObj, blueprint.sizeX - 1, it * (box.sizeY() - 1), k, BuildCraftFactory.frameBlock, 0);
+				blueprint.setSchematic(worldObj, 0, it * ((int) box.sizeY() - 1), k, BuildCraftFactory.frameBlock, 0);
+				blueprint.setSchematic(worldObj, blueprint.sizeX - 1, it * ((int) box.sizeY() - 1), k, BuildCraftFactory.frameBlock, 0);
 			}
 		}
 
@@ -652,7 +652,7 @@ public class TileQuarry extends TileBuildCraft implements IMachine, IPowerRecept
 			blueprint.setSchematic(worldObj, blueprint.sizeX - 1, h, blueprint.sizeZ - 1, BuildCraftFactory.frameBlock, 0);
 		}
 
-		blueprintBuilder = new BlueprintBuilder(blueprint, worldObj, box.xMin, yCoord, box.zMin, ForgeDirection.NORTH);
+		blueprintBuilder = new BlueprintBuilder(blueprint, worldObj, (int) box.xMin, yCoord, (int) box.zMin, ForgeDirection.NORTH);
 		blueprintIterator = blueprintBuilder.getBuilders().listIterator();
 	}
 
@@ -852,8 +852,8 @@ public class TileQuarry extends TileBuildCraft implements IMachine, IPowerRecept
 		chunks.add(quarryChunk);
 		ForgeChunkManager.forceChunk(ticket, quarryChunk);
 
-		for (int chunkX = box.xMin >> 4; chunkX <= box.xMax >> 4; chunkX++) {
-			for (int chunkZ = box.zMin >> 4; chunkZ <= box.zMax >> 4; chunkZ++) {
+		for (int chunkX = (int) box.xMin >> 4; chunkX <= (int) box.xMax >> 4; chunkX++) {
+			for (int chunkZ = (int) box.zMin >> 4; chunkZ <= (int) box.zMax >> 4; chunkZ++) {
 				ChunkCoordIntPair chunk = new ChunkCoordIntPair(chunkX, chunkZ);
 				ForgeChunkManager.forceChunk(ticket, chunk);
 				chunks.add(chunk);
